app/requirements.txt
Django==3.2
SQLAlchemy==1.4.11
Authlib==0.15.3
mysqlclient==2.0.3
djangorestframework==3.12.4
SQLAlchemy-serializer==1.3.4.4
django-cors-headers==3.7.0
drf-spectacular==0.15.1